+++ src/main/java/gift/global/domain/OAuthProvider.java
@@ -1,4 +1,4 @@
-package gift.oauth.business.dto;
+package gift.global.domain;
 
 public enum OAuthProvider {
     KAKAO
